% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGui.R
\name{runGui}
\alias{runGui}
\title{Launch the 'BsplineQuantReg' Shiny Interface}
\usage{
runGui()
}
\value{
Launches a Shiny application in the default browser.
}
\description{
Opens an interactive Shiny application for quantile regression
using B-splines with shape constraints.
}
\examples{
if (interactive()) {
  runGui()
}
}
