% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGui.R
\name{run_gui}
\alias{run_gui}
\title{Launch the 'BsplineQuantReg' Shiny Interface}
\usage{
run_gui(brow = TRUE, rstudio = FALSE)
}
\arguments{
\item{brow}{Boolean if TRUE launches the default browser in the system.}

\item{rstudio}{Boolean if TRUE launches the shiny browser of rstudio.}
}
\value{
Launches a Shiny application in the default browser.
}
\description{
Opens an interactive Shiny application for quantile regression
using B-splines with shape constraints.
}
\examples{
if (interactive()) {
  runGui()
}
}
